+++ UniFiTooling/UniFiTooling.psm1
@@ -158,6 +158,7 @@ function Add-HostsEntry
       Write-Verbose -Message 'Done'
    }
 }
+
 function ConvertFrom-UnixTimeStamp
 {
    <#
@@ -266,6 +267,7 @@ function ConvertFrom-UnixTimeStamp
       $Result
    }
 }
+
 function ConvertTo-UnixTimeStamp
 {
    <#
@@ -375,6 +377,7 @@ function ConvertTo-UnixTimeStamp
       $Result
    }
 }
+
 function Get-HostsFile
 {
    <#
@@ -460,6 +463,7 @@ function Get-HostsFile
       }
    }
 }
+
 function Get-UniFiConfig
 {
    <#
@@ -1194,6 +1198,7 @@ function Remove-HostsEntry
       Write-Verbose -Message 'Done'
    }
 }
+
 function Set-UniFiApiLoginBody
       {
          <#
@@ -2470,6 +2475,429 @@ function Get-UnifiNetworkList
    {
       # Dump the Result
       $Session.data
+
+      # Cleanup
+      $Session = $null
+
+      # Restore ProgressPreference
+      $ProgressPreference = $ExistingProgressPreference
+   }
+}
+
+function Get-UnifiSpeedTestResult
+{
+   <#
+         .SYNOPSIS
+         Get the UniFi Security Gateway (USG) Speed Test results
+
+         .DESCRIPTION
+         Get the UniFi Security Gateway (USG) Speed Test results
+
+         .PARAMETER Timeframe
+         Timeframe in hours, default is 24
+
+         .PARAMETER StartDate
+         Start date (valid Date String)
+         Default is now
+
+         .PARAMETER EndDate
+         End date (valid Date String), default is now minus 24 hours
+
+         .PARAMETER UnifiSite
+         UniFi Site as configured. The default is: default
+
+         .PARAMETER all
+         Get all existing Speed Test Results
+
+         .PARAMETER UniFiValues
+         Show results without modifications, like the UniFi Controller creates them
+
+         .EXAMPLE
+         PS C:\> Get-UnifiSpeedTestResult -all
+
+         Get all the UniFi Security Gateway (USG) Speed Test results
+
+         .EXAMPLE
+         PS C:\> Get-UnifiSpeedTestResult | Select-Object -Property *
+
+         Get the UniFi Security Gateway (USG) Speed Test results from the last 24 hours (default), returns all values
+
+         .EXAMPLE
+         PS C:\> Get-UnifiSpeedTestResult -UnifiSite 'Contoso'
+
+         Get the UniFi Security Gateway (USG) Speed Test results from the last 24 hours (default)
+
+         .EXAMPLE
+         PS C:\> Get-UnifiSpeedTestResult -Timeframe 48
+
+         Get the UniFi Security Gateway (USG) Speed Test results of the last 48 hours
+
+         .EXAMPLE
+         PS C:\> Get-UnifiSpeedTestResult -StartDate '1/16/2019 12:00 AM' -EndDate '1/16/2019 11:59:59 PM'
+
+         Get the UniFi Security Gateway (USG) Speed Test results for a given time/date
+         In the example, all results from 1/16/2019 (all day) will be returned
+
+         .NOTES
+         Initial version that makes it more human readable.
+         The filetring needs a few more tests
+
+         .LINK
+         Get-UniFiConfig
+
+         .LINK
+         Set-UniFiDefaultRequestHeader
+
+         .LINK
+         Invoke-UniFiApiLogin
+
+         .LINK
+         Invoke-RestMethod
+
+         .LINK
+         ConvertFrom-UnixTimeStamp
+
+         .LINK
+         ConvertTo-UnixTimeStamp
+   #>
+   [CmdletBinding(DefaultParameterSetName = 'DateSet',ConfirmImpact = 'None')]
+   [OutputType([psobject])]
+   param
+   (
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 0)]
+      [Alias('Start')]
+      [datetime]
+      $StartDate,
+      [Parameter(ParameterSetName = 'TimeFrameSet',
+            ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 1)]
+      [Alias('hours')]
+      [int]
+      $Timeframe,
+      [Parameter(ParameterSetName = 'DateSet',
+            ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 1)]
+      [ValidateNotNullOrEmpty()]
+      [datetime]
+      $EndDate = (Get-Date),
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 2)]
+      [ValidateNotNullOrEmpty()]
+      [Alias('Site')]
+      [string]
+      $UnifiSite = 'default',
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 3)]
+      [switch]
+      $all = $false,
+      [Parameter(ValueFromPipeline,
+            ValueFromPipelineByPropertyName,
+      Position = 4)]
+      [switch]
+      $UniFiValues = $false
+   )
+
+   begin
+   {
+      # Cleanup
+      $Session = $null
+
+      # Safe ProgressPreference and Setup SilentlyContinue for the function
+      $ExistingProgressPreference = ($ProgressPreference)
+      $ProgressPreference = 'SilentlyContinue'
+
+      #region CheckSession
+      if (-not (Get-UniFiIsAlive))
+      {
+         #region LoginCheckLoop
+         # TODO: Move to config
+         [int]$NumberOfRetries = '3'
+         [int]$RetryTimer = '5'
+         # Setup the Loop itself
+         $RetryLoop = $false
+         [int]$RetryCounter = '0'
+         # Original code/idea was by Thomas Maurer
+         do
+         {
+            try
+            {
+               # Try to Logout
+               try
+               {
+                  if (-not (Get-UniFiIsAlive))
+                  {
+                     Throw
+                  }
+               }
+               catch
+               {
+                  # We don't care about that
+                  Write-Verbose -Message 'Logout failed'
+               }
+
+               # Try a Session check (login is inherited here within the helper function)
+               if (-not (Get-UniFiIsAlive -ErrorAction Stop -WarningAction SilentlyContinue))
+               {
+                  Write-Error -Message 'Login failed' -ErrorAction Stop -Category AuthenticationError
+               }
+
+               # End the Loop
+               $RetryLoop = $true
+            }
+            catch
+            {
+               if ($RetryCounter -gt $NumberOfRetries)
+               {
+                  Write-Warning -Message ('Could still not login, after {0} retries.' -f $NumberOfRetries)
+
+                  # Stay in the Loop
+                  $RetryLoop = $true
+               }
+               else
+               {
+                  if ($RetryCounter -eq 0)
+                  {
+                     Write-Warning -Message ('Could not login! Retrying in {0} seconds.' -f $RetryTimer)
+                  }
+                  else
+                  {
+                     Write-Warning -Message ('Retry {0} of {1} failed. Retrying in {2} seconds.' -f $RetryCounter, $NumberOfRetries, $RetryTimer)
+                  }
+
+                  $null = (Start-Sleep -Seconds $RetryTimer)
+
+                  $RetryCounter = $RetryCounter + 1
+               }
+            }
+         }
+         While ($RetryLoop -eq $false)
+         #endregion LoginCheckLoop
+      }
+      #endregion CheckSession
+
+      #region ReCheckSession
+      if (-not ($RestSession))
+      {
+         # Restore ProgressPreference
+         $ProgressPreference = $ExistingProgressPreference
+
+         Write-Error -Message 'Unable to login! Check the connection to the controller, SSL certificates, and your credentials!' -ErrorAction Stop -Category AuthenticationError
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+      #endregion ReCheckSession
+
+      #region ConfigureDefaultDisplaySet
+      $defaultDisplaySet = 'time', 'download', 'upload', 'latency'
+
+      # Create the default property display set
+      $defaultDisplayPropertySet = (New-Object -TypeName System.Management.Automation.PSPropertySet -ArgumentList ('DefaultDisplayPropertySet', [string[]]$defaultDisplaySet))
+      $PSStandardMembers = [Management.Automation.PSMemberInfo[]]@($defaultDisplayPropertySet)
+      #endregion ConfigureDefaultDisplaySet
+
+      #region Filtering
+      switch ($PsCmdlet.ParameterSetName)
+      {
+         'TimeFrameSet'
+         {
+            Write-Verbose -Message 'TimeFrameSet'
+            if (-not ($StartDate))
+            {
+               if ($Timeframe)
+               {
+                  $StartDate = ((Get-Date).AddHours(-$Timeframe))
+               }
+               else
+               {
+                  $StartDate = ((Get-Date).AddDays(-1))
+               }
+            }
+
+            if (-not ($EndDate))
+            {
+               $EndDate = (Get-Date)
+            }
+         }
+         'DateSet'
+         {
+            Write-Verbose -Message 'DateSet'
+            if (-not ($StartDate))
+            {
+               $StartDate = ((Get-Date).AddDays(-1))
+            }
+
+            if (-not ($EndDate))
+            {
+               $EndDate = (Get-Date)
+            }
+         }
+      }
+
+      [string]$FilterStartDate = (ConvertTo-UnixTimestamp -Date $StartDate -Milliseconds)
+      [string]$FilterEndDate = (ConvertTo-UnixTimestamp -Date $EndDate -Milliseconds)
+
+      if ($all)
+      {
+         $FilterStartDate = $null
+         $FilterEndDate = $null
+      }
+      #endregion Filtering
+   }
+
+   process
+   {
+      try
+      {
+         Write-Verbose -Message 'Read the Config'
+         $null = (Get-UniFiConfig)
+
+         Write-Verbose -Message ('Certificate check - Should be {0}' -f $ApiSelfSignedCert)
+         [Net.ServicePointManager]::ServerCertificateValidationCallback = {
+            $ApiSelfSignedCert
+         }
+
+         Write-Verbose -Message 'Set the API Call default Header'
+         $null = (Set-UniFiDefaultRequestHeader)
+
+         Write-Verbose -Message 'Create the Request URI'
+         $ApiRequestUri = $ApiUri + 's/' + $UnifiSite + '/stat/report/archive.speedtest'
+         Write-Verbose -Message ('URI: {0}' -f $ApiRequestUri)
+
+         $Script:ApiRequestBodyInput = [PSCustomObject][ordered]@{
+            attrs = @(
+               'xput_download',
+               'xput_upload',
+               'latency',
+               'time'
+            )
+            start = $FilterStartDate
+            end   = $FilterEndDate
+         }
+
+         $paramConvertToJson = @{
+            InputObject   = $ApiRequestBodyInput
+            Depth         = 5
+            ErrorAction   = 'Stop'
+            WarningAction = 'SilentlyContinue'
+         }
+
+         $ApiRequestBodyInput = $null
+
+         $Script:ApiRequestBody = (ConvertTo-Json @paramConvertToJson)
+
+         Write-Verbose -Message 'Send the Request'
+
+         $paramInvokeRestMethod = @{
+            Method        = 'Post'
+            Uri           = $ApiRequestUri
+            Headers       = $RestHeader
+            Body          = $ApiRequestBody
+            ErrorAction   = 'SilentlyContinue'
+            WarningAction = 'SilentlyContinue'
+            WebSession    = $RestSession
+         }
+         $Session = (Invoke-RestMethod @paramInvokeRestMethod)
+
+         Write-Verbose -Message ('Session Info: {0}' -f $Session)
+      }
+      catch
+      {
+         # Try to Logout
+         try
+         {
+            $null = (Invoke-UniFiApiLogout -ErrorAction SilentlyContinue -WarningAction SilentlyContinue)
+         }
+         catch
+         {
+            # We don't care about that
+            Write-Verbose -Message 'Logout failed'
+         }
+
+         # Verbose stuff
+         $Script:line = $_.InvocationInfo.ScriptLineNumber
+         Write-Verbose -Message ('Error was in Line {0}' -f $line)
+         Write-Verbose -Message ('Error was {0}' -f $_)
+
+         # Error Message
+         Write-Error -Message 'Unable to get Firewall Groups' -ErrorAction Stop
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+      finally
+      {
+         # Reset the SSL Trust (make sure everything is back to default)
+         [Net.ServicePointManager]::ServerCertificateValidationCallback = $null
+      }
+
+      # check result
+      if ($Session.meta.rc -ne 'ok')
+      {
+         # Verbose stuff
+         $Script:line = $_.InvocationInfo.ScriptLineNumber
+         Write-Verbose -Message ('Error was in Line {0}' -f $line)
+         Write-Verbose -Message ('Error was {0}' -f $Session.meta.rc)
+
+         # Error Message
+         Write-Error -Message 'Unable to Login' -ErrorAction Stop
+
+         # Only here to catch a global ErrorAction overwrite
+         break
+      }
+
+      $Result = @()
+
+      foreach ($item in $Session.data)
+      {
+         $Object = $null
+         $Object = [PSCustomObject][ordered]@{
+            id       = $item._id
+            latency  = $item.latency
+            oid      = $item.oid
+            time     = if ($UniFiValues)
+            {
+               $item.time
+            }
+            else
+            {
+               ConvertFrom-UnixTimeStamp -TimeStamp ($item.time) -Milliseconds
+            }
+            download = if ($UniFiValues)
+            {
+               $item.xput_download
+            }
+            else
+            {
+               [math]::Round($item.xput_download,1)
+            }
+            upload   = if ($UniFiValues)
+            {
+               $item.xput_upload
+            }
+            else
+            {
+               [math]::Round($item.xput_upload,1)
+            }
+         }
+         $Result = $Result + $Object
+      }
+
+      # Give this object a unique typename
+      $Result.PSObject.TypeNames.Insert(0,'Speedtest.Result')
+      $Result | Add-Member MemberSet PSStandardMembers $PSStandardMembers
+   }
+
+   end
+   {
+      # Dump the Result
+      $Result
 
       # Cleanup
       $Session = $null